Plot updates and systematic on sigR measurement
increased error on fit to sigR to 0.005. To take into account the
systematic uncertainty from g1g2.

removed “Combined” from combined energy in all of the plots. We define
in the paper how “combined” energy is measured.

Cross checked that fastNEST also produces sigR=0.06*Ni even though the
input was Fr=0.0075 (to the modPoissRnd fluctuation on Ni)